pyproject.toml
[project]
name = "galaxy-publisher"
dynamic = ["version"]
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "pyjwt[crypto]>=2.9.0",
    "cryptography>=43.0.0",
    "pyyaml>=6.0.0",
]

[dependency-groups]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "coverage[toml]>=7.0.0",
    "ruff>=0.7.0",
    "mypy>=1.13.0",
    "ansible-core>=2.16.0",
    "python-multipart>=0.0.9",
    "types-PyYAML>=6.0.0",
    "respx>=0.21.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.dynamic]
version = {attr = "galaxy_publisher._version.__version__"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.12"
disallow_untyped_defs = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "integration: marks tests as integration tests (deselect with '-m \"not integration\"')",
]

[tool.coverage.run]
source = ["src"]
branch = true
parallel = true
concurrency = ["thread"]
sigterm = true

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.coverage.paths]
source = [
    "src/",
    "*/site-packages/",
]
